# Troupon.DealManagement.Api/swagger1.json
{
  "swagger": "2.0",
  "info": {
    "title": "Troupon Catalog API",
    "description": "A simple API to manage Catalog",
    "termsOfService": "https://example.com/terms",
    "contact": {
      "name": "Oualid Ktata",
      "url": "https://github.com/oualidktata/",
      "email": ""
    },
    "license": {
      "name": "Use under LICX",
      "url": "https://example.com/license"
    },
    "Version": "v1"
  },
  "paths": {
    "/api/Auth/token": {
      "post": {
        "tags": [
          "Auth"
        ],
        "produces": [
          "application/json",
          "application/xml",
          "text/plain"
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Server Error"
          },

          "502": {
            "description": "Server Error"
          },
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
    "/api/EngineTracker": {
      "post": {
        "tags": [
          "EngineTracker"
        ],
        "summary": "Gets enginetracker data",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SearchEngineTrackerInput"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/IncidentDto"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ValidationProblemDetails"
            }
          },
          "406": {
            "description": "Not Acceptable",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "409": {
            "description": "Conflict",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "202": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
    "/api/Incident": {
      "post": {
        "tags": [
          "Incident"
        ],
        "summary": "Gets incidents modified from a given modified date",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/IncidentSearchInput"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/IncidentDto"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ValidationProblemDetails"
            }
          },
          "406": {
            "description": "Not Acceptable",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "409": {
            "description": "Conflict",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "202": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
    "/api/Deals": {
      "get": {
        "tags": [
          "Troupon"
        ],
        "summary": "Gets all active Deals",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/DealDto"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ValidationProblemDetails"
            }
          },
          "406": {
            "description": "Not Acceptable",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "409": {
            "description": "Conflict",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "202": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Troupon"
        ],
        "summary": "Create a new Deal",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateDealModel"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Returned if the Deal was Created",
            "schema": {
              "$ref": "#/definitions/DealDto"
            }
          },
          "400": {
            "description": "Returned if the model couldn't be parsed or the customer couldn't be found",
            "schema": {
              "$ref": "#/definitions/ValidationProblemDetails"
            }
          },
          "406": {
            "description": "Returned if no response found with an acceptable format",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "409": {
            "description": "Conflict",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "422": {
            "description": "Returned when the validation failed",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
    "/api/Deals/{id}": {
      "get": {
        "tags": [
          "Troupon"
        ],
        "summary": "Gets a specific Deal",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/DealDto"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ValidationProblemDetails"
            }
          },
          "406": {
            "description": "Not Acceptable",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "409": {
            "description": "Conflict",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "202": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "ProblemDetails": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "SearchEngineTrackerInput": {
      "type": "object"
    },
    "IncidentDto": {
      "type": "object",
      "properties": {
        "date": {
          "type": "string"
        },
        "myProperty": {
          "type": "string"
        }
      }
    },
    "ValidationProblemDetails": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        },
        "errors": {
          "type": "object",
          "additionalProperties": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "readOnly": true
        }
      },
      "additionalProperties": {}
    },
    "IncidentSearchInput": {
      "required": [
        "businessUnit",
        "user"
      ],
      "type": "object",
      "properties": {
        "user": {
          "maxLength": 10,
          "type": "string"
        },
        "businessUnit": {
          "type": "string"
        }
      }
    },
    "DealDto": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "major": {
          "type": "string"
        },
        "minor": {
          "type": "string"
        },
        "patch": {
          "type": "string"
        }
      }
    },
    "CreateDealModel": {
      "type": "object",
      "properties": {
        "major": {
          "type": "string"
        },
        "minor": {
          "type": "string"
        },
        "patch": {
          "type": "string"
        }
      }
    }
  }
}